+++ server/src/constants.ts
@@ -1,0 +1,8 @@
+export const WIKI_LINKS = {
+    unitEventHandlers:
+        "https://community.bistudio.com/wiki/Arma_3:_Event_Handlers",
+    uiEventHandlers:
+        "https://community.bistudio.com/wiki/User_Interface_Event_Handlers",
+    commandsList:
+        "https://community.bistudio.com/wiki/Category:Scripting_Commands",
+};
