+++ server/src/lib/isNotNil.ts
@@ -1,0 +1,3 @@
+export const isNotNil = <T>(value: T | null | undefined): value is T => {
+    return value !== null && value !== undefined;
+};
